backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
django==4.2.5; python_version >= '3.8'
django-cors-headers==4.2.0; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
load-dotenv==0.1.0
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2023.3; sys_platform == 'win32'
